+++ src/backends/fst.rs
@@ -161,6 +161,7 @@ impl Reader {
                                 kind,
                                 definition_name,
                                 packing,
+                                is_hidden: false,
                             });
                             scope_ids.insert(key, id);
                             id